Rebuilt structure of the react app, I simplifide the flow of data in the app because my initial aproach was messed up.
The components display only according to the route thay recive and the ViewStore is updated only via those components on the route. Only the Menu component is subscribed to ViewStore and changes accordingly.

Also changed the components to be reusable insted of duplcating them for different routs.
Many bugs were fixed.

The Item view and edit view are still not working.